requirements.txt
requests>=2.32
beautifulsoup4>=4.12
python-dateutil>=2.9
icalendar>=5.0
